Separate packages neatly
