+++ app/src/main/res/layout/fragment_web_view.xml
@@ -6,13 +6,13 @@
 
         <variable
             name="viewModel"
-            type="com.theuhooi.uhooipicbook.modules.webview.viewmodels.WebViewViewModel" />
+            type="com.theuhooi.uhooipicbook.ui.webview.WebViewViewModel" />
     </data>
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".modules.webview.WebViewFragment">
+        tools:context=".ui.webview.WebViewFragment">
 
         <WebView
             android:id="@+id/webview"
