+++ app/src/main/res/layout/item_monster_list.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="monsterItem"
-            type="com.theuhooi.uhooipicbook.modules.monsterlist.entities.MonsterItem" />
+            type="com.theuhooi.uhooipicbook.data.monsters.impl.MonsterItem" />
     </data>
 
     <com.google.android.material.card.MaterialCardView
