+++ app/src/main/res/navigation/nav_graph.xml
@@ -11,7 +11,7 @@
 
         <fragment
             android:id="@+id/monster_list_fragment"
-            android:name="com.theuhooi.uhooipicbook.modules.monsterlist.MonsterListFragment"
+            android:name="com.theuhooi.uhooipicbook.ui.monsterlist.MonsterListFragment"
             android:label=""
             tools:layout="@layout/fragment_monster_list">
             <action
@@ -27,7 +27,7 @@
 
         <fragment
             android:id="@+id/monster_detail_fragment"
-            android:name="com.theuhooi.uhooipicbook.modules.monsterdetail.MonsterDetailFragment"
+            android:name="com.theuhooi.uhooipicbook.ui.monsterdetail.MonsterDetailFragment"
             android:label=""
             tools:layout="@layout/fragment_monster_detail">
             <argument
@@ -42,7 +42,7 @@
 
     <dialog
         android:id="@+id/dancing_monster_fragment"
-        android:name="com.theuhooi.uhooipicbook.modules.monsterdetail.dancingmonster.DancingMonsterFragment"
+        android:name="com.theuhooi.uhooipicbook.ui.dancingmonster.DancingMonsterFragment"
         android:label=""
         tools:layout="@layout/fragment_dancing_monster">
         <argument
@@ -52,7 +52,7 @@
 
     <fragment
         android:id="@+id/web_view_fragment"
-        android:name="com.theuhooi.uhooipicbook.modules.webview.WebViewFragment"
+        android:name="com.theuhooi.uhooipicbook.ui.webview.WebViewFragment"
         android:label=""
         tools:layout="@layout/fragment_web_view">
         <argument
